# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: table-order API
  version: 0.1.0
  description: |
    QR 테이블 오더 시스템 API 계약.
    설계 원본: docs/design.md (v0.2)

    - 소비자 API(`/t/{qr_token}/**`)는 인증 없이 경로의 `qr_token`으로 테이블 세션을 식별한다.
    - 관리자 API(`/admin/**`)는 `Authorization: Bearer <static-token>`으로 인증한다 (초기 버전, 사용자별 로그인 없음).

servers:
  - url: /

security:
  - {}

tags:
  - name: consumer
    description: 소비자(PWA) API — QR 토큰 기반
  - name: admin
    description: 주방/관리자 API — Android(태블릿·폰) 공용

paths:
  /t/{qrToken}:
    get:
      tags: [consumer]
      summary: 메뉴판 조회 (품절 반영)
      operationId: getMenuBoard
      parameters:
        - $ref: "#/components/parameters/QrToken"
      responses:
        "200":
          description: 메뉴판
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuBoard"
        "404":
          $ref: "#/components/responses/NotFound"

  /t/{qrToken}/orders:
    post:
      tags: [consumer]
      summary: 주문 생성 (자동 접수 → 즉시 RECEIVED)
      operationId: createConsumerOrder
      parameters:
        - $ref: "#/components/parameters/QrToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: 기존 주문 반환 (idempotency_key 재시도)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "201":
          description: 주문 생성됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 품절된 메뉴 포함 (sold_out_date 검증 실패)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /t/{qrToken}/session:
    get:
      tags: [consumer]
      summary: 내 테이블 주문 내역 조회
      operationId: getConsumerSession
      parameters:
        - $ref: "#/components/parameters/QrToken"
      responses:
        "200":
          description: 현재 OPEN 세션과 주문 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/orders:
    get:
      tags: [admin]
      summary: 대기열 조회 (폴링 겸용)
      operationId: listAdminOrders
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: 주문 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      tags: [admin]
      summary: 사장님 직접 입력 (source=COUNTER)
      operationId: createCounterOrder
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: 주문 생성됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/orders/{orderId}/ack:
    post:
      tags: [admin]
      summary: 표시 확인 (ACK, 가게 단위)
      operationId: ackOrder
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckRequest"
      responses:
        "200":
          description: ACK 처리됨
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/orders/{orderId}/complete:
    post:
      tags: [admin]
      summary: 조리완료 (주문 단위, 조건부 업데이트)
      operationId: completeOrder
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: 완료 처리됨
        "409":
          description: 이미 처리됨 (동시 조작 충돌)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/order-items/{orderItemId}/cancel:
    post:
      tags: [admin]
      summary: 품목 취소 (변심취소 또는 재고부족취소)
      operationId: cancelOrderItem
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderItemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [markSoldOut]
              properties:
                markSoldOut:
                  type: boolean
                  description: |
                    true = 재고부족취소: order_item 취소와 menu.sold_out_date 당일 설정을
                    한 트랜잭션으로 처리한다.
                    false = 변심취소: order_item만 취소하고 품절 처리는 하지 않는다.
      responses:
        "200":
          description: 취소됨 (markSoldOut 값에 따라 품절 처리 여부가 결정됨)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/menus/{menuId}/restore:
    post:
      tags: [admin]
      summary: 품절 해제
      operationId: restoreMenu
      security: [{ bearerAuth: [] }]
      parameters:
        - name: menuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 품절 해제됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/sessions/{sessionId}/close:
    post:
      tags: [admin]
      summary: 세션 정산/비우기
      operationId: closeSession
      security: [{ bearerAuth: [] }]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 세션 종료됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableSession"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/busy-mode:
    post:
      tags: [admin]
      summary: 바쁨 모드 on/off
      operationId: setBusyMode
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: 바쁨 모드 상태 변경됨

  /admin/devices:
    post:
      tags: [admin]
      summary: 기기 등록 (FCM 토큰, role)
      operationId: registerDevice
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: 기기 등록됨
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 관리자 API용 static token (초기 버전)

  parameters:
    QrToken:
      name: qrToken
      in: path
      required: true
      schema:
        type: string
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64

  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string

    OrderStatus:
      type: string
      enum: [RECEIVED, COMPLETED]

    OrderSource:
      type: string
      enum: [QR, COUNTER]

    OrderItemStatus:
      type: string
      enum: [ACTIVE, CANCELED]

    SessionStatus:
      type: string
      enum: [OPEN, CLOSED]

    DeviceRole:
      type: string
      enum: [PRIMARY, BACKUP]

    StoreTable:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        qrToken:
          type: string

    TableSession:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tableId:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/SessionStatus"
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true

    MenuCategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        sortOrder:
          type: integer

    Menu:
      type: object
      properties:
        id:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: integer
        sortOrder:
          type: integer
        isSelfService:
          type: boolean
        soldOutDate:
          type: string
          format: date
          nullable: true
          description: 품절된 영업일. null이면 판매 중.

    MenuBoard:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/MenuCategory"
        menus:
          type: array
          items:
            $ref: "#/components/schemas/Menu"

    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        menuId:
          type: integer
          format: int64
        menuName:
          type: string
          description: 주문 시점 스냅샷
        unitPrice:
          type: integer
          description: 주문 시점 스냅샷
        quantity:
          type: integer
        note:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderItemStatus"

    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sessionId:
          type: integer
          format: int64
        source:
          $ref: "#/components/schemas/OrderSource"
        status:
          $ref: "#/components/schemas/OrderStatus"
        idempotencyKey:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    CreateOrderRequest:
      type: object
      required: [idempotencyKey, items]
      properties:
        idempotencyKey:
          type: string
          format: uuid
        tableId:
          type: integer
          format: int64
          nullable: true
          description: |
            POST /t/{qrToken}/orders(소비자)에서는 무시된다 — 테이블은 경로의 qrToken으로 결정됨.
            POST /admin/orders(COUNTER)에서는 필수 — 어느 테이블의 구두 주문인지 지정해야 한다.
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [menuId, quantity]
            properties:
              menuId:
                type: integer
                format: int64
              quantity:
                type: integer
                minimum: 1
              note:
                type: string
                nullable: true

    SessionDetail:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/TableSession"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"

    AckRequest:
      type: object
      properties:
        deviceId:
          type: integer
          format: int64

    Device:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        role:
          $ref: "#/components/schemas/DeviceRole"
        fcmToken:
          type: string
        lastSeenAt:
          type: string
          format: date-time
          nullable: true

    RegisterDeviceRequest:
      type: object
      required: [name, role, fcmToken]
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/DeviceRole"
        fcmToken:
          type: string